Update bin.cpp
Resolved all conflicts.
This binary search algorithm is taking input from an input file. 

And the input is in the following format: 
1,2,3,4,5,6
1
2
3

First line is taking array input separated by commas.  
And for each of the next line there is present a number which is to be searched and returned with number of times comparisons were made .