Longest continious subset that adds up to zero
Given an array consisting of positive and negative integers, find the length of the longest continuous subset whose sum is zero.
To be done in O(n) time complexity. The use of hash-maps helps bring down the complexity to O(n).
Eg: Array size: 10 
    Array: 95 -97 -387 -435 -5 -70 897 127 23 284
    Then the length longest continious subset/sequence which sums up to zero is 5
    Since the longest continious subset which sums up to zero is -387 -435 -5 -70 897.

This program makes use of hashmaps and the STL is unordered_map.